macOS: initial support to macOS
This patch provides initial support to macOS.

It is required that `jq` is available and that `bash` version is 5.2 or
later. Both are available in [Homebrew](https://brew.sh), and, in case
of `bash`, the Homebrew `bin` path must come before the original system
path, as macOS (up to Sonoma 14.4) comes with Bash 3.x.